Post IRC logs to mediawiki.
This commit adds a simple script which posts irssi style irc logs to
mediawiki. This is useful for our private channel where I wanted the
logs to be searchable along side all our other html content.

There is a (poor) attempt to convert the logs into wiki style
formatting, but it could do with some work.